Initial commit
The purpose of this project is to write an extremely fast alignment
program that aligns a single query against an MSA. The use case would be
the alignment of a new viral strain against a previously prepared
selection of reference strains.

This commit includes a blosum90 matrix and influenza A protein files for
the H1-1A clade.