Toy with suffix trees
I have a rough notion of how to build an alignment program that writes
the reference MSA into a suffix tree and records the conserved "anchor"
points. Then the query would be aligned by matching to the anchor points
and then filling in the gaps. There are a many details to work out and
the algorithm will probably require a needlman-wunsch algorithm for
gap filling, so I think I will write the needleman-wunsch algorithm
first and come back to the fancy solution.