Implement an almost working alignment program
Almost. The algorithm seems to be failing to penalize trailing
non-matching sequences.

See the case: query=AAAGANDALF, subject=GANDALF

This should produce `GANDALF` with `AAA` inserted before position 0 in
the subject. But instead it produces `AANDALF`.

Also, the alignment is produced backwards. I just need to reverse it.
But I can't remember the magic word and I don't have an internet
connection.