Do not zero the last row
My idea was to make terminal gaps free. The implementation was grossly
incorrect (I honestly have no idea what I was thinking). The idea itself
is reasonable enough, but I'm not sure it is worth the complexity.

This fixes the weird failing case in the test suite.

The last consideration for alignment is how subject gaps should be
handled. There should be a positive score for aligned gaps. Gaps carry
information from the MSA the subject was pulled from.

For example, the alignment of `TTTW` against `T--TTW` should return
`T--TTW`, not `TTT--W`.